+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 - BREAKING feat: ESM
 - BREAKING fix: avoid setting `jsdoc` on `Program` or `Jsdoc` nodes themselves
+- BREAKING fix: avoid setting `jsdoc` on any descendants where the ancestor
+    can match
 - fix: avoid attaching `jsdoc` to `jsdoc` nodes
 - fix: clone `range` and `SourceLocation` rather than reusing
 - chore: fix linting